Move recursion validation test to nocover
