+++ hypothesis-python/tests/cover/test_deadline.py
@@ -155,76 +155,3 @@ def test_should_not_fail_deadline_due_to_gc():
         test()
     finally:
         gc.callbacks.remove(delay)
-
-
-@pytest.mark.skipif(
-    not hasattr(gc, "callbacks"), reason="PyPy has weird stack-limit behaviour"
-)
-def test_gc_hooks_do_not_cause_unraisable_recursionerror():
-    # We were concerned in #3979 that we might see bad results from a RecursionError
-    # inside the GC hook, if the stack was already deep and someone (e.g. Pytest)
-    # had installed a sys.unraisablehook which raises that later.
-
-    NUM_CYCLES = 10_000
-
-    def probe_depth():
-        try:
-            return probe_depth() + 1
-        except RecursionError:
-            return 0
-
-    def at_depth(depth, fn):
-        if depth <= 1:
-            return fn()
-        else:
-            # Recurse towards requested depth
-            return at_depth(depth - 1, fn)
-
-    def gen_cycles():
-        # We may be at the recursion limit, no free stack frames. Generate lots
-        # of reference cycles. Beware: there may not even be room for raising new
-        # exceptions here, anything will end up as a RecursionError.
-        for i in range(NUM_CYCLES):
-            a = [None]
-            b = [a]
-            a[0] = b
-
-    def gen_cycles_at_depth(depth, *, gc_disable):
-        try:
-            if gc_disable:
-                gc.disable()
-            at_depth(depth, gen_cycles)
-            if gc_disable:
-                assert gc.collect() >= 2 * NUM_CYCLES
-            else:
-                assert gc.collect() < 2 * NUM_CYCLES  # collection was triggered
-        finally:
-            gc.enable()
-
-    @given(st.booleans())
-    def inner_test(_):
-        max_depth = probe_depth()  # would have to be called twice on PyPy
-
-        # Lower the limit to where we can successfully generate cycles
-        while True:
-            try:
-                gen_cycles_at_depth(max_depth, gc_disable=True)
-            except RecursionError:
-                max_depth -= 1
-            else:
-                break
-
-        # Verify limits w/o any gc interfering
-        gen_cycles_at_depth(max_depth - 1, gc_disable=True)
-        # the previous line raises RecursionError on Pypy, while the next doesn't (!)
-        gen_cycles_at_depth(max_depth, gc_disable=True)
-        with pytest.raises(RecursionError):
-            gen_cycles_at_depth(max_depth + 1, gc_disable=True)
-
-        # Check that the limit is unchanged with gc enabled
-        gen_cycles_at_depth(max_depth - 1, gc_disable=False)
-        gen_cycles_at_depth(max_depth, gc_disable=False)
-        with pytest.raises(RecursionError):
-            gen_cycles_at_depth(max_depth + 1, gc_disable=False)
-
-    inner_test()
